Move couple of extension helper functions
